requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.101.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.22
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.20.0
aioredis==2.0.1
redis==4.6.0
sendgrid==6.10.0
APScheduler==3.10.3
requests==2.31.0